+++ src/features/account/hooks/details/useEmailForm.ts
@@ -4,10 +4,11 @@ import type { User } from '@supabase/supabase-js';
 import type { Dispatch, SetStateAction } from 'react';
 import { useMutation } from '@tanstack/react-query';
 
-import { DetailsEmailFormSchema } from '@schemas/account/detailsEmail.schema';
-import type { DetailsEmailForm } from '@schemas/account/detailsEmail.schema';
 import { updateAccountDetailsEmail } from '@server/account/updateEmail';
 import type { FormNotification } from '@typings/forms/notification';
+
+import { DetailsEmailFormSchema } from '../../schemas/detailsEmail.schema';
+import type { DetailsEmailForm } from '../../schemas/detailsEmail.schema';
 
 export default function useEmailForm(user: User | null, handleEmailEditShow: () => void, setEmailFormNotification: Dispatch<SetStateAction<FormNotification>>) {
   const {