+++ src/features/account/hooks/details/useNameForm.ts
@@ -4,10 +4,11 @@ import type { Dispatch, SetStateAction } from 'react';
 import type { User } from '@supabase/supabase-js';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { DetailsNameFormSchema } from '@schemas/account/detailsName.schema';
-import type { DetailsNameForm } from '@schemas/account/detailsName.schema';
 import { updateAccountDetailsName } from '@server/account/updateName';
 import type { FormNotification } from '@typings/forms/notification';
+
+import { DetailsNameFormSchema } from '../../schemas/detailsName.schema';
+import type { DetailsNameForm } from '../../schemas/detailsName.schema';
 
 export default function useNameForm(
   user: User | null,